[FIX] web client fix: crash when loading translation
Problem happens when the user from the website tries to load a translation,
it opens up a modal window without an action.  But the modal has a (hidden)
searchview which tries to load custom filters corresponding to that
non-existing action.

Another problem at the same time was that the view manager could not
display the breadcrumbs for a non-existing breadcrumb.