[REV] stock: revert of 0612486ebebdc1ee5d0fe1687d384353160e94d6
Reintroduction of field mes_type on product.template. This field was not
used in the code anymore, but we keep it for future use.

This field allows the user to express that there is no fixed conversion
rate between uom and uos.

(also very partial revert of 930f71d15ae656fffd66522ecdba54ffe45c8da1)