[FIX] fix reloading problem (crash) in web client
The previous viewmanager (before its refactoring in sept/oct 2014) was
ignoring some promises in its do_load_state method.  The refactoring
actually fixed that, but of course, some other code was dependent on
the broken implementation.  This commit removes the use of promises in
do_load_state, which is functionally equivalent to the previous code.

As a result, it relies on some subtle timing of the creation and
initialization of various views.  It is somewhat fragile, but it is
actually what the former code did, only in a more explicit way.