[FIX] web: fix searchview menus height
first, I want to say that I'm not proud of this.  Let's just say that it
is better than the previous code (of which I'm not proud either).

The issue is that the sub menus (filters/groupbys/favorites) aren't very
good when their height is too big, and they're hidden below the screen
without scrollbar.  Previous code tried clumsily to compute the correct
max height in js, but it was only done in filters menu, and did not
always work.  This code do the same in css.  This would be good except
that it is just an estimation: calc(100vh - 125px).  It assumes that
the header will approximately take 120px, and size the max-height
accordingly.