[FIX] web_calendar: error in case of creation with quick_create disabled
Broken in 4296a4f94c1 porting of quick create to Bootstrap modals: in
case quick_create is disabled, the creation of the quick_create dialog
is bypassed and so is defining this.$input, but focus() calls still
happen and blow up.

Create a dummy $input in the constructor so it's alway possible to call
this.$input.focus().