[FIX] web: session requires a name to properly work
For historic reasons, the session had a name (hardcoded to "instance0"),
used in at least one place: the get and set cookie methods.

The idea was most likely to allow multiple instances of the web client
to coexist peacefully.  This does not work at all at least for the last
versions, and the new module system is a better way to do it anyway.

However, before removing completely the prefix in cookies, we should
actually check if no external code depends on it.