[FIX] Web: view.js: rpc call to load sidebar content
this.options.$sidebar usage has been changed at rev. 82d9c21b71e87f47876cb7198897164bddb282e0
and a flag this.options.sidebar was introduced ($sidebar now simply
represents a jQuery element where the sidebar has to be rendered,
whereas sidebar represents whether or not a sidebar has to be rendered).

The old variable $sidebar was still used for the rpc to get the content
of the sidebar stored server-side, resulting in a sidebar with missing
content.