[FIX] website: fixed element should have a position
Otherwise, it is fixed according to the other elements, here, the wrapwrap element.
So if we add a margin to wrapwrap, the margin is also added to the navbar, which
means it is not really fixed on the top.
This fix avoids this issue so the navbar is always fixed at the top of the page.