[FIX] web: don't crash when switching views in firefox
For some weird reasons, IE and Chrome (or jquery?) bind the global
variable 'event' to the last triggered event (I guess).  Of course,
our code depended on that behavior, but Firefox does not like it and
crashes when we try to read the 'event' variable.  This fix simply gets
the variable as first argument.