[FIX] tools: yaml_import: filter returned values of _create_record.
Commit a33753351ab95c4b67bde51c09d5e041318e6be4 added default values for all fields present
in the view to emulate. However some fields present in the view are not stored and should
not be sent to create. This bug appears with function field in new API that are not stored.
Using the Yaml import generated warning, for example for fields like alias_domain that is
a not stored computed field.