[IMP] translate: translation optionally wrap translated terms inside a <span>
Note: We do not translate `<a>` elements inline.  Translating `<a>` elements is
triggering issues in the website editor tool: menu items are screwed up.  The
menu items are handled by bootstrap, which expects menu items to be `<a>`
elements inside `<li>` inside `<ul>` elements.  The introduction of a `<span>`
in the middle of this hierarchy breaks its functionality.  Therefore we decided
to not translate `<a>` elements inline.

When invoking `read()` with `'website_editor`' in the context, missing
translations are automatically created, so that translated terms are always
wrapped as `<span data-oe-translation-id="...">...</span>`.