[FIX] web: prevent useless rpc in one2many
In some cases (for example, some popup dialogs with a form view), the
one2many field is destroyed but the async code still tries to append
a view manager to itself.  The result is that the start method of the
view manager is run, resulting in at least an useless fieldviewget.