[IMP] fields: when computing a field, do not recompute other stored computed fields
In the process of determining a stored computed field's value, the field could
not be read from the database if the environment was in "draft" mode.  The idea
behind this condition was that during an onchange, no computed field should be
read from the database.  We fix the condition by checking whether we are in
"onchange" mode instead of "draft" mode.

The change may be critical for performance: one of the purposes of storing a
computed field is to avoid recomputing it when not necessary.  Now a computed
field can read the value of a stored computed field from the database.

This issue was spotted by Cédric Snauwaert (csn), who came with a concrete
example were a stored computed field was recomputed without apparent reason.