[FIX] web: load sidebar after the loads of view fields
and the render of the buttons.

The sibar menu can depends on the current view and fields
states.
We therefore have to wait for them to be loaded before loading
the sidebar.

Before this rev., loading a report view (in the reporting menu),
leaded to a crash, due to this revision: ac92551406c333dc6829126e4c2fc6f0f1c613f9.

opw-640409