[FIX] event: Attendees stat button in event form/kanban
The stat button count included only reserved seats,
while, when clicking on the button, all attendees
were displayed: draft, confirmed, canceled, done.

Usability speaking, this isn't right.

Besides, once the attendee actually attended the event,
it wasn't anymore included in the count, which is not logical
either: by "Confirmed", you expect the reserved attendees
and the attendees already there.

In addition, on the kanban view, the link said "confirmed" attendees
as well, but redirected to the list with all attendees.
The "expected attendees" value on the kanban was along a "report"
link, bringing to the report where only confirmed seats are displayed.

Kind of non-sense.

Finally, the decision made:
 - For an event, the more natural count that you want to see first
 is the "expected" attendees, including draft, reserved and attended
 attendees.
 - Both buttons, on the form view and the kanban view now includes
 draft, open and done attendees, and redirect to the list view
 displaying only these attendees (canceled excluded).
 - On the kanban view, you have a count more including the confirmed
 attendees, including the open and the done attendees.

opw-642134