[IMP] stock: Improve lots with serial numbers, define lot tracking by picking type and convert stock with lot False to a lot when necessary
Instead of having only outgoing, manufacturing, incoming, all... traceability on the product,
we can define on the product if there is no tracking, lot tracking or unique serial number tracking.
On the picking type we can define whether we will only create lots for the picking type or use existing lots.

The lots are required when the product and the picking type demand it.

The option for serial number will check that each piece (quantity 1) has its
own serial number.

A lot splitting wizard is added on the pack operations in the picking form.

We want to be able to handle the case where we receive a lot of iPads e.g. and we
only assign a serial number when the product is delivered.  For this, when the product is
delivered, we see that we only have products with no lots.  When we see that we take something
without lot for something that needed one, we assign the lot to that piece of stock.

For negative quants: a positive quant with no lot can reconcile with a negative quant with/without lot,
a positive quant with lot, will preferably reconcile with a negative quant with the same lot, afterwards
with that without lot.