[FIX] web_tests,test_new_api: act_window for new design compatibility
The widget_x2many tour uses the discussion model to test one2many
and many2many widgets. The tour clicks on an item of the menu which
opens the discussion model with the discussion_form_2 view (in create mode).
Then, it clicks on the view_switcher to display the list view, and clicks
on an item of the list to open a previously created record.

In the new design, there is no more view_switcher in form view. This flow
is thus unfeasible. This solution changes the menu action so that it
directly opens the model with the list view, on which we can click on the
record to display in form view.