[FIX] mail: discard changes confirm dialog
This change is required for new design compatibility because when
toggling the app switcher, the whole content is detached. The app
switcher is toggled once the deferred def is resolved. If def is
resolved before the dialog is closed, the dialog is detached as well,
and re-attached when clicking on another apps. With this rev., def is
resolved once the dialog is closed.

Also fixes the on_cancel() call which wasn't done at the right place.