[FIX] sale_stock: restore clean inheritance
Inheritance with position=attributes has been introduced at a6870dd6ae but was
screwed up during conflict resolution of merges and lost some inheritance
or got the one on product_id to be duplicated.

Set inheritance with position=attributes everywhere.
Add comments to explain when product_uos should be kept or not.