[REF] web: ControlPanel
Adds the option clear so that, if set to false, only some elements of the
control panel can be updated while keeping the other ones unchanged.
The breadcrumbs can then be updated by simply calling update_control_panel(),
instead of triggering a specific event to do it.

A bit of documentation and refactoring as well, mainly for the
breadcrumbs to be handled as other control panel's elements.