[IMP] web_kanban: scroll to the top on page changed
This commit only concerns enterprise edition and is absolutely useless in the
community one.

When clicking on the pager to skip to next/previous page of a kanban view, the
expected behavior is that the scroll position go back to the top. When editing
the pager (e.g. changing the limit), however, the scroll position should
not change.