[IMP] fields: optimize `setup_base()` for regular (non-related) fields
Split the methods `_setup_regular` and `_setup_related` in `base` and `full`
parts, such that the `base` part is done in `setup_base()`, and the `full` part
is done in `setup_full()`.

This optimizes the setup of models during the installation of modules: regular
fields are not set up again from scratch.  Only related fields (which are less
common) are always set up from scratch.