[IMP] hr_timesheet: get rid of the "Service" product
The existence of the "Service" product caused issues when users would
eg. remove this product.

We allow user to create timsheets even if they dont have an employee associated
or if the product is missing on the employee. Those timesheets can still be
invoiced using the force product field of the wizard.

product_id is left blank if it cannot be deduced from the associated employee
and the amount is zero.

account_id is left blank we it not required anymore

journal_id is set the first general journal found.