[FIX] web: no selection widget in search with many2one fields
In previous releases, the search view ignored the selection widget when
the field is a many2one.  This behaviour is justifield by the fact that
the selection widget just doesn't work, and the default search input
actually does work.

The search view was refactored extensively in late 2014, and this
behaviour was lost.  As a result, every many2one with a "selection"
widget stopped working (for example, warehouse/products, a search for
"you" does not give anything on "location" or on "warehouse").  This
commits simply fix the issue by ignoring the selection widget in search
views (with many2one fields).