[FIX] web_kanban: better behaviour when grouping
* fix a crash when grouping on non-many2one fields (due to the
group.values object being not initialized)
* display correct label when grouping on selection fields

We should have a better way to deal with fields differences.  There is
this weird concept called "abstraction", which could be applied and
would allow everyone to focus on their business and not on dealing with
this, but that would probably not fit with our codebase.