[IMP] project: removed members.
Members were not used in project, except for the now removed faces library.
Statistics can still be computed based on task or issues responsibles
instead of adding all members of a project.

Indeed you could have members that did not actually work in the project
making it difficult to differentiate with followers. Now the concept
of members is removed. Followers is sufficient for communication.