[FIX] hr_timesheet: magic values when is_timesheet
Timesheets entries could only be created with the help of the magic
default_is_timesheet=1 set in the context, thus in views.

To allow manual creation or import of timesheets by setting is_timesheet=True
directly, the values of invisible but required fields will be computed at
create insted of default_get.