[FIX] lunch: fixed cash move balance stat button
The awful code of 43f843e didn't even work when
there really wasn't any cash move for a user for
which you didn't have the rights to read the moves.

Because there wasn't any move, there wasn't any
access error either.

We finally decided to hard-code the rule that
allows you to read the balance or not, which
renders it dependent of the actual access rules
of the system.