[IMP] website_sale : add rating into ecommerce
When posting a comment, the user can rate the product using rating star widget. The rating card is display on the product page with average and repartition.
Add a related field 'website_published' on rating in website_sale for performance issue.

Part of this commit could be moved in 'website_rating', if the feature is required in other modules. Since it is only used in ecommerce, the code belong to website_sale.