[FIX] point_of_sale: take quantity into account when calculating taxes
Fixes a bug introduced in 5014a36b8b28fcf202eb162761ac722c6b254b64.
Before this, creating an order with two of the same products would
result in a total price to pay which would only include the tax of one
product. This would cause point of sale orders in the backend which
weren't fully paid.