[FIX] remove explicit naming of m2m indexes
Postgres has a limit of 63 characters on identifiers, because of the way
m2m index names are built the limit could relatively easily breached
leading to a truncation of the index name, and in some cases collisions
between the name of the two index.

<name> in CREATE INDEX is optional since Postgres 9.0, use this feature
instead of generating index names, should avoid any possible collision
on m2m index names.

closes #7620
fixes #2780 (master)