[FIX] analytic: methods used in default attribute should return a RecordSet
If a method used as a default returns an id instead of a RecordSet,
it could fail in 'convert_to_write' (Many2One field) because it will try
to return "value.id". If "value" is an integer, it fails.

Also changed the method to a lambda expression and removed the
'_default_user' method because it's not used anywhere.