[FIX] l10n_be: multilang translations settings
The translations inside the i18n/ folder are the translations published on
Transifex. It will contains the translatable "general" terms (menus, error
messages, actions...).

The account, taxes, journal names are made translatable thanks to the
l10n_multlang module but it should only be translated in the `spoken_languages`
values (nl_BE for l10n_be). These translations are present in i18n_extra folder
which is not published on Transifex.

Regenerate the translations correctly and populate the nl_BE.po file with
previous translations.