[FIX] account, ir.sequence: compatibility for accounting sequences using today's date.
In saas-6, the `ir_sequence_date` key in the context
has been introduced, to use this date instead of the
current date for sequences including date variables
(Year, month, day, ...).

So, when an invoice is created for a previous fiscal
year / period, the sequence is built according
to the date of this fiscal year / period.

e.g., for an invoice sequence containing the year,
for an invoice created in 2015, but which
is for the 2014 fiscal year, the sequence
contains 2014 instead of 2015.

Nevertheless, this leads to a retro-compatibility issue:
the impossibiliy to use the current date for your
invoices sequences.
Indeed, some users want to include the current month
and day in their invoices sequences,
e.g. xxx-20150922.

This revision brings the possibility to use both
possible dates, the beginning of the date range
or the current date.

opw-649487