[IMP] point_of_sale: fiscal position per pos_order
Allows a user to set multiple fiscal positions per pos_config. Orders
created on that config can then be assigned one of these fiscal
positions. In the POS this can be done through a button that will appear
above the numpad when fiscal positions are set on the pos_config.