[FIX] test_new_api: rewrite `test_related` to avoid changing binary fields
In the former test, a stored binary field was recomputed.  Changing its value
meant changing an attachment, and the latter was deleting a file in the
filestore, because that file was not used by another attachment.  The problem
was: that operation is rolled back from the database (test transactions are not
committed), but not from the filestore!