[FIX] point_of_sale: set taxes when creating pos_order from form view
When creating pos_orders from the form view, tax_ids on the
pos_order_lines were never set.

This was kind of fixed in b69f5cd4c1955c7022bc0ce76eb585eb740fbfec. But
in there it was in onchange_qty, because the way the fiscal position was
taken into account required it to be. Because we take the fiscal
position into account in a different way now we can assign the taxes in
onchange_product_id which makes a bit more sense.