[IMP] auth_crypt: update to Passlib 1.6
Debian Jessie provides Passlib 1.6 so we're not bound to 1.5 anymore.

Use 1.6's "deprecated=auto" feature: will automatically mark any hash
other than the default as "to update" rather than having to list all of
the supported hashes to deprecate (and remember to update in lockstep
with the schemes list).

When not specified explicitly, the default hash is the first one in the
schemes list.