[FIX] website_quote: remove loading attribute if missing data
In commit b5030fe6, a loading attribute is set on any submit button. The
problem is that this attribute is the first thing to be set. Therefore,
we cannot prevent its propagation.

Since Bootstrap sets this 'loading' attribute in a setTimeout, we need
to remove it in a setTimeout as well.