[FIX] web_editor: null range and remember_selection
Tests for null range were not done in #8845, so in given case (where the
focus is not in the editor) we could get an issue.

The code for remembering the selection when clicking the editor
was made useful in 281a043d096, but there was an edge case when
the editor itself was in a modal: then the selection would be lost.

This commit change the code as following:

1) restore the selected range if:

- the current selected range is not in the editable of the active editor
- there is a corresponding saved selected range or we are in an embeded
  editor (and not in air mode as when editing the website)

2) save the selected range if:

- we don't restore the selection,
- the current selected range is in the active editor editable area