[FIX] website_mail,mass_mailing: better support for outlook
This fix is about several issue of office oulook:

- a fixed table width must be in pixel, using 100% of parent div width
  doesn't work,

- a background-color on a div is not used, the background color should
  be on the table inside this div using the attribute "bgcolor",

- the only way to center a table is to have a `<center>` tag arround it,
  or have an element arround it with "text-align: center",

- having a background image doesn't seem to be possible.

The two first are straightforward, but for the third:

- in non-outlook: all the snippet are centered (since they all have
  "margin: auto" which is valid outside outlook),
- in outlook: only a small part of the snippet with text-align:center
  or `<center>` arround them are centered.

To escape a bigger code change, this commit unify the outlook
behaviour by having all of the snippet left aligned in outlook (they are
still middle centered in other application).

For the background image, we need a fallback bgcolor and thus we need to
have the background on the table element. There was also an issue
between relative and absolute url for gmail.

closes #9019
opw-650217