[FIX] website_sale: currency in line with pricelist selected
Before this fix, inconsistencies arise in the prices displayed in a
multi-currency environment. For example, let's consider these 2
pricelists:
- Public Pricelist, in USD
- Benelux Pricelist, in EUR with 10 % discount

When the Benelux Pricelist is selected, the strikethrough price is in
USD, the price is in EUR, and the currency displayed is $. The reason
is that the pricelist used is stored in request.session, but the
templates are still using the pricelist linked to the partner.

We link website.pricelist_id to the currency stored in request.session,
which is more logical. We also use the currency related to this
pricelist in the templates, which avoids the inconsistencies mentioned.