[FIX] point_of_sale: exclude rescue sessions when getting current session
Rescue sessions have been added with revision
7fd12b6dcd62af46e8e0bcc1799c120aa20a011f

With this feature, this is possible to have multiple
pos sessions opened for a same user, if there are
rescue sessions.

`get_current_session` rely on the fact only one pos session
can be opened per user.

Rescue sessions are now excluded to get the current session, so:
 - at most one session can be found
 - rescue sessions cannot be resumed.

opw-652857