[FIX] account: partner read access in multi-companies
The below revision:
3daa47303178a3296f16ac90a9e03d078bab7c04

Makes the multi-companies record rule disabled by default,
meaning the user can see contacts from other company than
his current company.

Nevertheless, to read the computed field `currency_id` on
the `res.partner` model, the user need the read access
on the contact `company_id` field, that the user does not have
if the partner company is not a child of the user current company.

opw-653103