[FIX] web: no accesskey for x2m pager
The fix in 7e46824 solved the issue for pager of some X2Many widget x2m
which would shadow the accesskey to go to previous/next page of the top
view's  current model.

But in the case a web.Pager (used in community for a X2Many kanban
widget) the fix had no effect.

opw-652888