[FIX] web_editor: escape for xml in translation as done in ir.ui.view
Currently server side there is two differents source for html content
displayed on a qweb rendered page:

1) ir.ui.view arch value

are stored as xml and are unescaped before being displayed, so we have
to escape the text nodes before saving.

e.g:

* &nbsp; should be stored as &amp;nbsp; since it is an HTML entity but
  not a defined XML entity,
* &amp; should be stored as &amp;amp; since when unescaping it, it
  would become & which is invalid in HTML,
* &lt; should be stored as &amp;lt; since when unescaping it, it would
  become `<` so HTML text node could become HTML element node.

The tag themselves can also be escaped but it often has the same result:
&lt;em&;gt;blah&lt;/em&gt; is the same as <em>blah</em> when unescaping
is applied to both.

But it is required is some instance, e.g <!DOCTYPE html> unescaped would
be invalid xml.

2) openerp.fields.Html value

are stored directly in HTML formatting and should, conversely, not be
escaped before being stored.

Thus when saving an ir.ui.view part modified thanks to the website
editor, an escaping of text node content is carried out to take this
into account, but the same was currently not done for the v9.0 new
translation system.

This commit apply the same logic when saving ir.ui.view as when saving
ir.translation values.