[FIX] sale, website_sale: send email at sale order confirmation in ecommerce
When confirming a sale order through the ecommerce
(at payment, most of the time),
the confirmation email should be sent at confirmation.

This is the subject of the revision
809e1a60813e0468ffc3f52f866119bb8cfe3a58
that was lost with revision
34751e14caf5ff8345efea03cfe56a2f5726ee04

At the moment, the email was sent only for
orders not being automatically confirmed
(transfers, for instance).

opw-653800