[FIX] website_sale: attribute selection unreversible
On a list of product, the attribue list (can be activated via the
customize menu) only contains the attribute pertaining to current
filtered product list.

Also, all the values of a the attribute are showed, even if choosing a
value will result in an empty product list.

It could lead to usability issue in some situations.

e.g

1) the current showed products number 2 with an attribute color,
2) we select "Red" which isn't a value these 2 products are declined in,
3) the current show products is now empty, but we can't un-select the
   "Red" color since "Color" is not an attribute of any of those 0
   product.

This commit solves this issue by displaying the currently selected
attributes if no product is displayed.

opw-653594