[FIX] stock: when moving lots it should also take into account pack operation info
When moving lots, we should also take into account the information on the
pack operation: specific location, package to take it from and owner.

Also, in the get_preferred_domain for the pack operations, when not giving any pack,
we should check the quant is not in a pack.  It is also better to copy the list
to avoid that the constraints added to the list would get into the next call.