[FIX] mail: close livechat conversations without unsubscribing
The problem occured when the user in the backend (client action) closed the
livechat before the user in the frontend, because we performed the wrong RPC,
and the backend user was removed from the partners following the discussion.
Thus, the messages sent from that point by the frontend user were never
received.

When closing a DM or livechat, we should call 'channel_pin', when unsubscribing
from a channel, we should call 'action_unfollow'.