[FIX] web: editable list validation with several pages
This revision is related to 5fa1f197a2b2b76c45fe1b8a2f3a03d524f88b57.

`dataset.cache` contains an item for each records, even
if the records hasn't been loaded in the cache. If a record
has not been loaded, only the record id is available, not
the other values. Thefore, this isn't possible to validate
the record values of records that have not been loaded
in the dataset cache.

opw-654013