[FIX] stock: don't prepare pack operations when not everything is reserved and delivery method is all at once
When all the products need to be reserved before doing the picking,
the pack operations were not shown, but still created.

This is not necessary (and did not work as it tried to reserve according to
existing pack operations), as you will need to reserve again when there is
sufficient stock to put the picking in available state and then
the pack operations will be generated and made visible.