[IMP] web: graph legend tweaks when many item
The legend of the graph when there was too much items in it could
prevent user to see a part or even the whole graph.

On non-mobile device, move the Pie Chart legend to the right, thus if it
overflows the char is still visible and we can toggle the seen top
legend item.

Otherway, the legend on top is hidden when :

- there is more than 25 legend items and the device has width < 768
- there is more than 50 legend items and the device has 768 <= width < 992
- there is more than 75 legend items and the device has 992 <= width < 1200
- there is more than 100 legend items and the device has 1200 <= width

Since this is based only on the number of legend items but legend items
could have variable name lengths, there will still be edge case where
the legend obstruct the graphical data display.

In this situation the current way to cope with this further filter the
displayed data.

opw-653445