[FIX] point_of_sale: wait on invoicing to finish before printing receipt
After this commit was introduced b428e43392bd2881e823e0b80a0191eeadd31fd5 , When you click on validate with invoicing enabled without customer,
the popup for select a customer open and immediately we switch to the  receipt screen, that gets rid of popup.
In this commit we print the receipt after invoice process done in backend when validate an order with invoicing enabled so that it's work like before.