[FIX] point_of_sale: don't ignore iface_tax_included
In 9baa63067a69dbebe20dcda3b4acb0dcf43eade9 the concept of
iface_tax_included was introduced. It was not actually used for the
display price though. It looks like a mistake during merging or
something because it adds the code responsible for dealing with the
option after a return statement in get_display_price().

opw-657983