[FIX] mail: message not found
message_ids contains the ids of all the messages. However, a maximum of
100 messages are fetched when a page is open, and therefore some
messages might not be found in the JS cache. This leads to
message.is_needaction throwing an exception.

This fix returns True if the message is not found. It means that the
message will be set as 'read' even if it is not displayed.

opw-657896