[FIX] mail: thread: oe_mail_expand: execute handler once
The read_more button displayed in long messages is an <a> inside a <span>, both
having classname oe_mail_expand, and an handler in bound on that classname
since rev. f933060e. This handler was executed twice when clicking on that
button, toggling twice the message's bodies long and short, thus doing nothing.