[FIX] website_mail_channel: wrong override
message_get_email_values is declared as receiving cr, uid, id. However this method
is an override of a multi method. A decorator has been added to tell that id is
actually a list of ids and avoid list encapsulation issues with the result. This
way the method parameters are not changed.

[CLEAN] mail: removed some unused imports, by the way