[FIX] base: guess_mimetype
When receiving an email containing two attachments(.docx, .odt) with the chatter
the icons displayed for the two attachments were the icon of a zip file instead
of the icon of a document file.

ps: the dictionnary "mapping" with the mimetypes and the signatures has been
changed in a tuple of tuples to keep the order of the mimetypes because
all the document files (.docx, .dot) begin by 'PK\x03\x04' and all the
zip files by 'PK'.

opw:658706