[FIX] base: fix SignedIntConverter weight
This allow to keep route with an SignedIntConverter with more
priority than a StringConverter.

The default wight is 100, so force the weight to 50 (default intConverter)
was also a fix. But we decide to rewrite the SignedIntConverter based on the
intConverter of WerkZeug.

So now, a controller '/module/<string:xmlid>', will be matched after
'/module/<int:id>". (Same behaviour that before the commit d13474894aaf95080e9a604de082008cb7d609c1)