[FIX] point_of_sale: make tendered amount adhere to currency formatting
Part of the number formatting overhaul done in
754ca789df779ba15011261fe4d7630f97e8f95f,
4b003afd0ab125b0296de8939626ae889ccdab47,
261ccf7c56062930b1d712ae860fe01dd8718968 and
d5430f6d63318cc8d7c97c269d4ffb363573a85e. The only place left where we
didn't correctly format the numbers was in the tendered column on
the payment screen. This fixes that by not directly displaying the
inputbuffer anymore. Instead we pass the payment line amount through
format_currency_no_symbol() and display that instead. We also give a
default amount of 0 to parse_value so it doesn't return false for an
empty string.