[IMP] website_slides: improve not pretty thumbnail
In 9.0 image_medium and image_thumb route are used to display thumbnail
of website_slides document.

These in 9.0 will scale the image, but also crop it so it has a good
aspect ratio.

In saas-6 this croping feature is not present and these thumbnail will
be scaled in diverse ugly way.

To improve a little the situation, this commit use some CSS (currently
not supported only by internet explorer) which give a similar aspect as
9.0's got.

opw-658957