[FIX] website_quote:
This revision is related to 0795e985e27e3d8fa5a73897d234e23c95de7b30

The payment acquirers must be fetched if the quote
needs a payment.

Besides, as explained in the above revision, the state
`manual` no longer exists sales orders, and therefore
this must no longer be used to determine if a quote
needs a payment or not.

opw-660575