[FIX] web: views in action dialog
Opening a view (kanban or list) in a dialog thanks to a target=new
action did not work anymore (only in community).

The cause was the 'search_view' action flag which was set to false
in case of a popup (preventing the view to search the records and so
to display them). This commit sets this flag to true if the popup
has to display a list/kanban view.

Backport of aab enterprise work.