[FIX] point_of_sale: open POS dashboard when no sessions on /pos/web
When you go to /pos/web and there is no open session that belongs to the
current user the controller redirected to an old action called
action_pos_session_opening that was removed in
eec760bbcaf841f3cfe7646e134d3e3fc3ae7ca3. This replaces that action
with action_client_pos_menu which just redirects to the POS dashboard.