[FIX] website_mail: safari mail issue for font awesome
In current safari (tested on 8 and 9) it seems things like:

content: "\f000";

are transformed when getting them javascriptally to:

content: \f000;

which seems to make no sense with CSS specifications but tests verified
this happened.

This broke sending mass mailing in some instance, which is solved by
this commit.

opw-665496