[FIX] Remove timezoning in stock_calendar
Commit 64e1c28e3 introduced timezoning in resource_calendar in resource already.
This broke the stock_calendar module, but as such, the timezoning in the module
is not necessary anymore.  All dates calculated from the times in the calendar
are calculated back to UTC and the results are without timezone in UTC.  That way,
timezones are also not needed for calculating the leaves.