[FIX] purchase: prevent write on the related `stock.location`
`default_location_dest_id_usage` on `purchase.order` is
used only for an attrs domain on the field
`dest_address_id` in the purchase order form.

There is therefore no worries to set it as readonly,
it won't break the feature he has been added for.

Besides, this will prevent the useless write of the
same field value on the related `stock.location`
when saving the purchase order.

This is important to prevent the useless write
on `stock.location`, as users might have the rights
to create/edit purchase orders without having
the write access to the stock locations.

opw-660736