[FIX] sale: unlink SO and procurement
When a SO is confirmed, cancelled then set to draft, the SO is still
linked to the pickings and the procurements which were created before
cancellation. This should not be the case, since a draft SO is not
supposed to be linked to any procurement.

The fix resets the procurement group of the SO, and resets the
procurements linked to a SO line.

Fixes #10277