[FIX] product: pricelist applied on product categories
This is related to revision f85830cbcd6377af1d1fe2a9b700c8b9e1c304eb

The above revision introduces a regression: the ability
to configure pricelist items applied on product categories.

If you configured a pricelist with items applied
on products categories, then the first item with a product
category matched all products, even if the products were
not in the according category.

This is because the condition checking that the product
is within the pricelist item product category has been
completely removed, by the above commit.

opw-666059