[FIX] mail: fix issue with disappearing filters in some cases
The client action had a weird issue: going back from a breadcrumb to the
client action made the searchview (filters/favorites) buttons disappear.  This
is caused by a weird behavior of the control panel.  This fix is the safest
for now, but working on the control panel behavior would be a better long term
solution.