[IMP] point_of_sale: limit categories height to 60%
When categories on the point of sale take too much height, the product
and overflown categories are not visible.

In those instance Odoo wizards would usually:

- use an option to have small categories without pictures,

- have a category "all" with all the product,

- do a custom change to reduce the size of categories or add a scrollbar

This commit add the last customization in stable.

If the category block is bigger than 60% of the page height, the
categories are cropped and a scrollbar is added.

closes #10588

opw-660383