[FIX] website_event_sale: Tour test, the expected amount is 5,500
The amount was 3,8560.96 because the Benelux pricelist was
applied by default, due to a regression introduced by
6b22a1ba59e0e53cc3d2f6f0ee3abb1d2e7a6888

while it should have been the public pricelist loaded
by default.

Both the benelux and the public pricelists were available,
but it should be the public pricelist chosen by default,
as it's the default website pricelist.

See 06760743aa3329ff9f8dc540221405a595979563 for more
explanations.