[FIX] website_forum: a partner can have multiple users.
A partner can be associated to multiple users.

The condition checking the partner
user has enough karma should therefore
handle it. The previous condition expected
a singleton (only one user for the partner).

With the condition before this revision,
a partner with multiple users could
simply not post new questions on the forum.

opw-667894