[FIX] product: product price assumed currency
Before this revision, with a configuration setup
like this:
 - Company A - Currency EUR
 - Company B - Currency USD
 - Public Pricelist- Currency EUR
 - Product A - Company A - Price 100 (EUR)
 - Product B - No company - Price 100

When a user set in Company B browses the `/shop`,
the pricelist applied is the public pricelist,
and, therefore, the prices are shown in EUR
(nothing wrong so far),
but the currency of the price set on the products
were assumed in USD, the user company currency, and,
therefore, the rate conversion from USD to EUR
was applied.

The product A, configured with a sale price of 100€
in the company A, was therefore displayed to 92€
(100€ * the USD-EUR rate).

The product B, without any company set, as well.

Meaning, two users from two different companies using the same
pricelist see two different prices in the same currency,
because their companies do not use the same currency.
This doesn't make any sense.

With this revision,
when setuping a sale price on a product, the currency
of this price is assumed in the currency of the company
set on the product.

If there is no company set on the product (the company
is not required), the currency assumed is then the currency
of the main company.

opw-666059