[FIX] ir_sequence: `range_` actually uses the range start date
In a sequence, `range_year` is supposed to be the year
of the range beginning date:

```
When subsequences per date range are used,
you can prefix variables with 'range_' to use
the beginning of the range instead of the current date,
e.g. %(range_year)s instead of %(year)s.
```

Therefore, for a sequence with ranges configured as:
 - 20150701 to 20160630
 - 20160701 to 20170630

An invoice confirmed on 20160101 must have `2015`
as year when using `range_year`, and not `2016`.

If you would like the invoice effective date,
you must then use `year`, instead.

We add the `current_%` keys as well,
in case some users absolutely want the
`now` date instead of the effective date,
but it doesn't make much sense in the context
of accounting.

Besides, outside the context of the accounting,
`year`, `month`, `date` etc is the `now` date.

opw-649487