[FIX] product_visible_discount: trigger property on overriden `product_uom_change`
This revision is related to dc76a7cfbccdd0569217cd5cb90b5904b2e72deb

When overriding an onchange method, the `@api.change` property,
that decides on which fields the onchange method must be triggered,
should be the same on the overriding and the overriden method.

Otherwise, the base onchange won't be triggered on the fields
it expected to.

opw-668049