[FIX] calendar: To avoid duplicate contacts
Before the fix ae6d0b77b8849318d4327ada1bb35cbf2277e81a, the duplicate
records were erased thanks to the index used in the dictionary(id of the partner).
After this fix, to avoid this this problem, the label of each record is used as
key in the dictionary.