[FIX] mail: announcement bar URl displayed on print
On printing, the announcement bar URL was displayed
in the printing result.
e.g.
`https://services.openerp.com/openerp-enterprise/ab/register/...`
was displayed at the top of the print.

This is related to the revision da2131b8fe4a242248253746cd152275f3cdd237.

This is due to this simple typo, the according CSS rule was
never applied because of it.