[FIX] crm: merge of partners twice in the action menu
The link to open the merging partners wizard was duplicated
in the action menu:
 - Once as "Deduplicate Contact",
 - Once as "Merge selected contacts".

This is due to the revision f36a093,
converting the
`<record model="ir.actions.act_window"...`
to an
`<act_window`
which creates an action binding automatically.

No action binding must be created for this window
action, as there is already another one,
the link to this wizard is already in the
partners action menu, thanks to the `act_window`
`action_partner_merge`.

Besides, the `'default_state': 'option'` seems useless,
as this is the default value set in the model directly.
On the opposite, removing the `active_test` (as done in
the revision mentioned above) could be armful. We
therefore re-set it.

opw-669688