[FIX] purchase: purchase stat button list domain
From revision a3d3e67,
the purchase stat button is only on product template form,
and no longer on the product variant form.

The domain of the list that this purchase button
opens has not been adapted according to this change,
the domain was still applied as coming from a product
variant form rather than a product template form.

As a result, it opened the purchases of products variants
having as ID the product template ID from which you came
from before hitting this button.

opw-669618