[FIX] mrp: _action_explode
Before the fix:

Create a product with a BoM type kit(composed with 5 products) and price=120
Select real price costing method for the product and each component
Select perpetual inventory valuation  for the product and each component
Create a PO with the product
Receive the products

When checking the journal of entries, 5 entries have been created with an
amount of 120.

After the fix:

When checking the journal of entries, 5 entries will be created with an
amount equal of the standard price of each component of the kit.

opw:670239