[FIX] base: shield embedded view from action 'create'
With `1bcfb0e` if a window action has `'create': False` in its context,
creating a record by one of the action view is disabled.

But this is also applied to eventual embedded views which is not wanted.

With this commit, a `'create': False` in the context is only applied on
root view model and not on embedded ones.

closes #11207
opw-660607