[IMP] product: Pricelist creation on Company creation if new currency
When working in multi-company, and each company using its own currency,
a pricelist for the second company should be created. For instance, our
default company is in Belgium -> € and there is a default pricelist in €.
Then you create a second company in US -> $. It is impossible for the us
company to sell in its own country because there is no $ pricelist.

When creating the second company, if the currency is different than the
first company one, the default pricelist in the currency should be created.

In all the cases, a ir.property should be created to link the
'property_product_pricelist' on the 'res.partner' to the new company.

Furthemore, a small usability improvement. The currency should be
activated before creating the company. So we change the label to indicate
that behavior and make the action in target new to prevent the creation flow
to be broken