[FIX] i18n: escaping of incorrect html in help messages
The field help on an ir.actions.act_window is a translatable html field.
When a value is set, it is processed by the `html_sanitize` method to make sure
only valid HTML content is present in the database.
In sanitization, xml line break tag `<br/>` is converted into html tag `<br>`.

Before revision db4041ff, `xml_translate` method was using XML parser for html
fields. Line break was then exported as `<br/>`.
Since this revision, it is using a proper html parser and line break is exported
the same way as it is stored in the database, with a `<br>`.

When translating html value, the content is split per meaningful block and
recomposed based on the msgsrc value. Due to the change of parser, translations
exported before db4041ff contains `<br/>` tag inside the msgstr and could not
match any block in splitted content (which contains `<br>`).

Correct the .pot with valid tag.
Do not change the xml content of help message as it should still be XML inside
the data files.

opw 671477