[FIX] product_extended: _get_variant_count
Due to the _set_product_template_field function, it is not possible
to set the standard price on a product template with more than one variant.

Due to the _compute_product_template_field function, it is not possible
to read the standard price on a product with more than one variant.

The field "standard_price" is in readonly to avoid some problems if the costing
method is in real time.

Closes: #11308

opw:671013