[FIX] barcodes: first qualify the quantity event before handling it
Introduced by 530fe35bebc80114a45512095a277bf7c8dbbd22. Before doing
anything we should first ensure that the quantity keypress event targets
body (and that it is a number, but that's less important). Before this
we didn't do that and thus the _display_no_edit_mode_warning was given
when people where just typing in the associated thread of a picking.

opw-672350