[FIX] warning: onchange_partner_id
When the warning module is installed, `onchange_picking_type` is not
called anymore.

When the method `onchange_picking_type` was introduced, it was not taken
into account that the warning module would change the onchange property
on the `partner_id` field thanks to the method `onchange_partner_id`.

opw-672091
Fixes #11342