[IMP] web_editor: inline buttons shown over topbar
Before 9.0 web_editor separation in 510e48d, some style would make
inline button be visible over topbar.

This commit reintroduce that and also make the margin over the inline
editor content a little bigger (so the inline buttons on the top of a
content will cover a little less the topbar).

closes #11526
opw-671649