[FIX] purchase, purchase_requisition: make_po
When a product is set up to generate a purchase tender instead of a
purchase order, the associated procurement is set in exception.

This is because the result of `make_po` is returned by the method `_run`,
and the result of `_run` is used to determine if the procurement is
running or in exception. Since a purchase tender does not generate the PO,
the method `make_po` returns an empty list, and therefore the procurement
is considered as in exception.

The fix is to return the procurement_id instead if a purchase order or a
purchase tender has been created.

opw-672170