[FIX] mrp_repair: tax calculation and subtotal
Following commit d36be782, the the taxes on the fees are not calculated
correctly. Moreover, setting the subtotal in the onchange is misleading
since it will be set to zero if the line is not invoiced. We just set
the price to zero by default, and the click on 'update' will set it
correctly afterwards.

opw-673984