[FIX] account: partner_id field in account.move
From 8ab34f93b64a08a4e1b407660ced1c5a8094f564

The original issue is the following:
- Go to the journal item creation
- Choose a partner
- Add a line => the partner on the journal item is removed

This is because `partner_id` on the account move is a related to
`partner_id` on the account move lines. This is likely to cause errors
during the manual encoding of a journal item, since the relation will
link the partner of the account move to the partner of the first account
move line only.

opw:672354