[FIX] calendar: Partial revert of 08d3ef00986ef89b4d9e8f9ac562ec9f96604b71
The above revision allow to edit an event with a start date
greater than the stop date.

This issue was solved by the revision
6c78541978336ae8ad4dc43f7051c19bf1ad4f7d
that the above revision reverts.

The reason why the constraint on
`event.start` and `event.stop` was replaced
by a constraint on
`event.start_datetime` and `event.stop_datetime`
and
`event.start_date` and `event.stop_date` are explained
in 6c78541978336ae8ad4dc43f7051c19bf1ad4f7d:
Basically, `start` and `stop` are computed fields,
that are not recomputed in time, before the constraint is checked.

Besides, the revision 08d3ef00986ef89b4d9e8f9ac562ec9f96604b71
re-set the constraint on `start` and `stop`, but forget
to replace the fields in the constraint definition:
```
(_check_closing_date, 'Error ! End date cannot be set before start date.',
['start_datetime', 'stop_datetime', 'start_date', 'stop_date'])
```

opw-673297