[FIX] point_of_sale,website_sale: form view layout (avatar)
Classname 'oe_right' was overriden by 'oe_avatar' before rev. 127d2ae, and thus
the images were displayed on the left of the sheet. Since rev. 127d2ae, the
'oe_right' takes the priority over 'oe_avatar' as it should. This commit
restores the layout of these two form views as it was before 127d2ae.