[IMP] point_of_sale: Stat button to invoice on pos_order view
Purpose:
In POS order form view the 'Re-Print' button is placed improperly.
There should be a way where invoice can be linked to POS without this button..

Removed the reprint button from the form view of POS order.
And add the invoice state button so invoice can be linked to POS.