[FIX] account: SO reference in register payment wzd
Follow up of fix fb0d969

This previous fix removed the reference field when creating an
invoice from a sales order since this field was for Supplier
Invoices (and is not shown for Customer Invoices).

While this males sense, it had the side effect of removing the
default value for the "Register Payment" wizard.

To keep this wizard populated with the most sensible data,
the default is now set to reference or name or number (in that
order).

opw 676261