[IMP] mail: chatter: better sync between chatter and form view
When switching between records, the displayed thread was still the one of the
previous records until the messages of the current record were loaded, which
could be confusing.

Moreover, when quickly browsing records using the pager, it could happen that
the resulting displayed thread wasn't the one of the displayed record (if
we visited a record with a long thread history, the RPC to load that history
might return after the one of the current record).

This commit aims to always synchronize the displayed thread with the current
record, by
  - directly hide the previous thread when browsing records
  - displaying a spinner after 1s to indicate that the thread is loading and
    (possibly) not empty
  - ensure to only render the thread of the last visited record