[IMP] account: Improve graph proportions in accounting dashboard
========
Purpose
========

The graphs on the accounting dashboard give the impression that we had $0 in the bank because the graph is at the bottom of the kanban kard.

==========
Specification
==========

The bottom should be zero, the top should be the maximum of the visible period, and the rest of the graph should be proportionate.