[IMP] mrp_operation: prevent resize on Gantt view
Resizing a workcenter line is confusing to the user since a resize
only changes the `date_planned_end` and recalculates the `date_planned`.
However, a user would expect that a resize changes the duration, which is
not the case.

To avoid the confusion, it is not allowed to resize the tasks anymore.

opw-676762