[FIX] account: duplicate journal should not use the same bank account
There is several override of create and write that assumes that a bank account
is present on only one account (see 0864be4).
It is very hard to change or remove the account bank once linked to a journal
(see cafe619d) so it should not be set in the first place when duplicating a
journal.

opw-676374