Revert "[FIX] res.currency: allow duplication"
This reverts commit d780f9476d43eb4fd25c852dd4e691002f4a6dcf.

Did not work due to the size=3 on name field that strips the code to the first
three letters only (removing the " (copy)" part).
Copying a currency has a few business cases as the rates are not copied.
As can not increase the size of a field in stable, remove the method that had no
effect.

Fixes #11036