[FIX] web,base: m2o empty name results
M2O use the autocomplete lib to display its content. The problem is
that if a result is empty (or only composed of whitespaces), a
divider is displayed in the dropdown.

Commit 98cc0e15086e67e1edbf5368f3e21f0b672e2ba6 was fixing this by only displaying non-empty
name results... but it did not work as many models implement the
name_search method the wrong way.

The new solution is to add a default display for record values which
are empty strings, at the right places in the webclient components.