[FIX] product: pricelist item sorting
The issue occurs in the following case:
- Configure 2 product categories A and B such as A is the parent of B
- Create a pricelist
- Create a first pricelist item which applies to category A
- Create a second pricelist item which applies to category B (with the
  same configuration regarding the minimum quantity)
- Create a product for category B

Before this fix, we cannot ensure that the pricelist item related to
category B will be applied, since we do not order the pricelist items by
category.

Thanks to this fix, we make sure to select the category from the more
specific to the less specific.

opw-678633