[FIX] res_currency: company rates subselect
The sub-select used to fetch companies rates
was wrong in the case rates with no company set
were entered before rates for specific companies.

e.g., for USD, the below rates:
Date        Rate        Company
2016-02-13  0.721400
2016-02-25  0.738800    1
2016-02-26  0.740000    1

the sub-select returned
Rate        Start       End
0.721400    2016-02-13
0.738800    2016-02-25  2016-02-26
0.740000    2016-02-26

while it was expected to return
Rate        Start        End
0.721400    2016-02-13  2016-02-25
0.738800    2016-02-25  2016-02-26
0.740000    2016-02-26

as the rate 0.721400 ended on 2016-02-25,
since a new rate specific to this company 1 was
added at that date.

opw-678726