[FIX] crm : apparent bug in handle_partner_assignation
Starting from c637e58dfe295d38c75c8e7c53aa56454ef6346d if a lead is
already assigned to a partner instead of returning the already assigned
partner we ignore that, possibly reassign the lead and return either the
new assignment or ``False`` (if not in action=create and no partner_id
was provided).

This seems like an error which passed through and was never noticed as
the method isn't called much and only from wizards.

Either way, revert that. And remove the comment which apparently was
never updated.