[FIX] point_of_sale: ensure that there is always an order creation_date
Introduced by e9d38bbc546219b0b536a123759f244a92f4fd13. This commit made
sense but there are certain POS modules that send the order to the
backend before validation and thus depend on there always being a
creation date, even if the order has not been validated yet.

This ensures that we don't end up with an undefined creation_date in the
backend.

opw-680590