[ADD] BoM price report
In v8, we had a BoM price report in xsl format which was removed in v9.

As it is still needed, we added a report in the new qweb format, but this
time it is on the BoM instead of the product variant.

It also makes sense to have a similar report for v10 as the cost structure report
on the mo is for the real cost and not the standard cost.