[FIX] product: do not display state field
The field was removed from the form view at 51d85ad8 during a refactoring but
was forgotten in the tree view.
It makes no sense to display a field that the user can not change.

The same field was already invisible on the product.product tree view since
43977de.

Make it invisible instead of removing it to avoid breaking potential xpaths.

Fixes #12418