[8.0][FIX][event_type] Allow the user to choose typeless events.
The requirement for somebody to choose a ticket product should be that it is an event, not that it has an event type attached, mostly when `event_type_id` is not a required field.

In fact, the domain should include products that are tickets and have no type or the same type as the current event, but that would alter the database layout, so I'm not patching that for this version.