[FIX] website_quote: Notification 'viewed by customer' should be received by the responsible only
Currently some customers receive the automatic notification 'Quotation
viewed by customer'. This should not be the case as this notifciation
is purely internal and should only be sent to the quote responsible.

Notification is now a note with the responsible being directly in the
recipients.

Conflicts:
	addons/website_quote/controllers/main.py

Backport of 4c4e5a3e10498bc0e282d7fb323db4773084c903
opw-687957