[FIX] stock: no rereservation needed in a lot of cases and make sure the moves passed to action_done are those from the recomputation
It is possible (in very rare cases) on rereservation that the new reservation will trigger other stock moves
than the first reservation.  In order to be sure to send the correct stock moves involved to done, it is better
to do the rereservation before the splitting of the moves on the total picking.

Also rereserve is optimized as it should only reserve for what is given by the pack operations,
as the backorder will be reserved later on anyways.