[ADD] website_project_timesheet: Timesheet on Tasks in website portal.
Show the timesheets (account.analytic.line) recorded on a task in the website
portal.

website_project_issue_sheet: once a timesheet can be read, use sudo to read the
name of the responsible as portal users might not have the right to read users.

Original authors: Vipul Bhatt, Florian Wintjens