[FIX] account: unnecessary default currency
A mechanism already exists to select the appropriate currency and was
introduced in commit 6a0cd97. However, it requires a purchase journal in
the currency of the PO to work.

Note that this commit is likely to not work correctly since the method
`_onchange_journal_id` is triggered after `_default_currency`.
Therefore, even if the currency is correctly set in `_default_currency`,
it is changed to the journal currency right after.

This reverts commit b596889574ddfd9401fcb59d3afb8a8591c5330c.