[FIX] *: html fields sanitization
Since commit d3c8cb151302dc397d4a65a873f6ea3ac32e5e19 in saas-10, the html fields are sanitized in
strict mode by default. This breaks all html fields behaviour when it
is editable in the website editor.

This commit deactivate the sanitization for such HTML fields (like it
was already the case for blogs contents or products descriptions for
example). A more specific sanitization will be implemented in future
versions.