[FIX] web: default many2many tags color
When there is no `color` field on the model
e.g. On invoice line taxes
The color class to apply on the badge was set to
`o_tag_color_undefined`
which doesn't exist. It must be set
from `o_tag_color_0` to `o_tag_color_10`
in order have a color.

Because of this, the taxes in the invoices lines
were not visible, because they were written in white
on a white background.

This has been broken by the below revision:
8854542a6107665a358fe26f9825e664f993e06a
removing the or statement (`||`) to set a default
color for models not having a `color` field.

We choose to set `10` as default value, so the color
remains the same compared to the previous release
(light grey). However, if the new behavior chosen
by the usability is to have these in black, it should then
be changed to `0` again.