[FIX] website: remove useless extra trigger in banner tour
The tip does not have to wait for this extra_trigger to show up.
Also, waiting for it seems to induce some bugs on chrome under
certain screen sizes, so best to avoid them.

Alse reduce the size of this tip as it is too big for small screen
sizes.