[FIX] account: remove related payments upon statement line cancellation.
Since commit 471daf304e6eddeeb4d941bc2c08b00c5ceef9a9, each statement line creates a payment entry if it creates a move. So the code to cancel a statement line must be adapted to keep the old behavior: remove the payment if it was created by the statement line processing; keep the payment if it was registered before the statement line.