[FIX] web: session_info: `is_admin` check
The tour mechanism displays a tour if the `session.is_admin` key
is set to True. For this key to bet set to True, you need to be
in the system group.

Default options adds the system group to any created user, however as
there aren't any ir.model.access rule defined on the `web_tour.tour`
model, when you try to login as this created user, you get a traceback.

As the only way to reach a `web_tour.tour` record is to have the
superuser_id, update the `is_admin` condition accordingly.