[FIX] product_visible_discount, sale: compute the right unit price
When the module product_visible_discount is installed, the field
'discount_policy' allows to display the discount in the sale order or
allows to consider the discount as include in the price_unit.

The field price on product.product always compute the price according to
the pricelist then it always includes the discount in the price.

The field lst_price on product.product never compute the price according to
the pricelist then it always gives the sale price.

According to the discount_policy, the price or the lst_price must be taken.

opw:683771