[FIX] calendar: use correct bootstrap for invitations
Commit fddff5d8d4d160bd2d3e8eab47afa0a38ae0e006 introduced a new template for meeting invitation page.
This template used bootstrap row's without any container which added
an ugly useless scrollbar at the bottom.

This commit uses valid bootstrap by simply removing the bootstrap
column layout (as there is no column but only centered content) and
using a simple container. Also, wraps the table in a .table-responsive
div so that the template is mobile-friendly.
Also fixes a typo in the template name.