[FIX] website_sale: force recompute fiscal position
Fiscal position was setted by onchange before commit #1d65f5f
Now, we need to call the onchange_shipping_id to set the fiscal position.
In case where the owner of the sale order is the public user, instead to
use the fiscal position from public user, we try to guess it with the
geolocation.