[FIX] web_editor: paragraph dropdown style
Before commit a8ee18e172d406e846764b1dd79c65ee270ca7f4, the paragraph alignment dropdown menu was
too large for its content, so it was ugly. After that commit, it became
too small for its content, which is even more ugly.

This fix sets the width of this specific dropdown so that its content
fills exactly in (as with original summernote).