[FIX] web_editor: paragraph dropdown style
Before commit 6ef338019339774dd6980704e5802d6b0e4cd795, the paragraph alignment dropdown menu was
too large for its content, so it was ugly. After that commit, it became
too small for its content, which is even more ugly.

This fix sets the width of this specific dropdown so that its content
fills exactly in (as with original summernote).