[FIX] account: don't use abs price in invoice report
The field price_subtotal_signed on an invoice line contains the line
subtotal:

- untouched if the invoice is of in_refund or out_invoice type,
- negated if the invoice is of in_invoice or out_refund type

At one point in time (not found), commit 1d28977c using absolute value
of price_subtotal was introduced because price_subtotal_signed sometimes
had the wrong sign.

But this is source of an issue since it is possible to have an invoice with
two lines such as : 100€ and -20€, the report would then display 120, whilst
80 should have appeared.

This commit remove the use of absolute value, and if some old report data
(pre saas-6 migration with wrong migration script) are reported wrong
for an instance, it should be corrected manually.

opw-683952
closes #13118