Revert "[IMP] project: tour mini improvement"
This reverts commit 584746ca03f885769022f827dfcb645b8a78a7fa.

The commit failed the tour test since there is invisible
".o_kanban_ghost" cards after the last. Since the created new
project is already sorted in the right location, we don't know
if it will be the first or the last.

But if it is the first, we know we will see the help tip.