[FIX] qweb: fix ir.qweb.field.contact
Before this commit when you edit a contact with the editor, you didn't see
the selected record before the save.

Now, you can edit a contact with the html editor.
We move the hacked field self (on res.partner) from website_partner to base.

2 bugs:
   field and record was reversed in call
   need to pass a browseRecord and not a dict, because we uses .env on it

This hack allow to use widget='contact' on the browse record
itself and not on a specific field from the browse record

Eg:
    <div t-field="partner.self" t-options='{
                 "widget": "contact",
                 "fields": ["address", "phone"]
    }'/>