[IMP] website_sale : add sort by
Now you can sort products by lst_price, name, ... via the UI

In some case the price list or b2b/b2c mode can altered the price
and the sort on lst_price can be sometimes wrong.
Eg: if price is displayed vat in (b2c), and mix rate of 6% and 21%, the sort
will be done on the htva price.