[IMP] base: limitate access to config settings to superuser or group_system members
Currently members of group_erp_manager (Access Rights) and group_system
(Settings) have theoritically access to config settings. However only
group_system members have access to the settings through the interface
as all menus are restricted to them. Moreover after changing a config settings
a check on ir.actions.todo is performed. This model is limited to group_system.

It is therefore not possible for group_erp_manager members to execute config
settings. This commit therefore change the check performed in config settings
to ensure user is superuser or a member of group_system to avoid unwanted
access rights errors further in the process.

[IMP] sale_stock: remove now unnecessary check