[FIX] website, website_slides: slug of a non-existing record
When you execute the slug method on a virtual record (not yet
saved, virtualid/newid) the slugs obviously fails as it needs
an id and a display_name.

This issue happens when there's a depends on a function field
using the slug method: just after the create and during the
onchanges, they'll raise.

This patch makes the slug method returns None in these case,
anyway the depends are re-triggered after the write and at
this time they have an id and everything is working fine.