[FIX] theme_*: restore category to theme
category was changed to "Website" so it'd appear under the website
category in the store, but it needs to be "Theme" otherwise it doesn't
appear in the "Choose your theme" menu after installing website.

This is a backport of 10.0 https://github.com/odoo/odoo/commit/7a1bac16

Without extra themes, the "Choose your theme" would be empty.

fixes #21898
opw-1802936
closes #24042