[FIX] mass_mailing, website_mass_mailing: "group_website_popup_on_exit" group
This group was defined in the mass_mailing module but is only working when
you install website_mass_mailing. What this group does is allowing to insert
a newsletter snippet in the page, thus it has nothing to do in the mass_mailing
module which doesn't depend on website. Also, the snippet itself was already
defined in website_mass_mailing.

Moved all the things to website_mass_mailing.