[FIX] account: fix exchange rate entries
When a reconciliation creating an exchange rate entry was done, the newly account_move_line corresponding to the exchange entry was not given a full_reconcile_id.

Also when doing a full reconciliation in different currency, full_reconcile_id was not set on account_partial_reconcile object resulting in problem when trying to unreconcile entries.

This is a backward fix of commit: 621fd8808d693ac293300fedc50ed9b69765966c and a470bb50411b3bb7118b44a1c4b257947dca40d9