[IMP] website_blog: use valid website snippet as default content for blog post
Adapt the default content of blog post to be a valid website snippet.
Previous content was not entirely possible to remove (there still was
a remaining .container) and was not easily customizable (user had to
remove it if he wants something great).

Note: maybe we should adapt the snippets so that they are callable with
custom content or do something like what has been done for default
content for job positions (see commit b760a35caffcf62bf44355bbc5d30908e47bc6b0).