[FIX] product: remove duplicated field product.template form view
When the purchase model is installed, there are two checkbox with
label 'Can be Purchased'. This error is due the rev https://github.com/odoo/odoo/commit/40ad5248bc43550771902e30aa6a2a8587ea8889
where we moved purchase related fields from module product to an
inherited view in purchase module.

To fix this, we removed the 'purchase_ok' field that was left in
the product_template view.