[IMP] account: clean create method from mandatory variable
Avoid setting a value for a related field as it will trigger onchange update.
This code was relevant when the date was different on the move and the line but
is no longer needed.

Remove all the `if vals.get(mandatory_field)` to simplify this method and remove
dead code or condition always true.