[FIX] mrp: fix BOM smart button for product variants
Do not force the `search_default_product_id`to the current
product variant, as the domain just above could return a
result, a BOM which is set for the template of this variant
but not for the variant istself, that this filter would hide.

This is following the revision
52cba1b56325f101ef159648d5d1dd3b974576f0

Because of this, a product variant smart button could
display "1 BOM", but when clicking on it, none would be displayed,
which is not user-friendly at all.

opw-687912