[FIX] website_sale: toggle shipping state display
In d055b321 the code displaying state of a given country was improved
for some incompatibilities.

But, if when setting the billing informations we choose to create a new
shipping address, the shipping state* would not be displayed even when it
should be (currently, if the shipping country is Australia or U.S.).

opw-687948

*a state being a subdivision of a country