[FIX] service: initialize_db: correctly set country
When we are creating a database with a country code from the database manager,
the country is not set in the company, so when we install the accounting app, there
is no chart of account installed by default. We have this bug because
we are trying to put a country id in a field that should receive a
'res.country' record.

To fix this bug, we change the call of function read_search by a search
that will return a recordset.

Bug introduced in rev https://github.com/odoo/odoo/commit/aeab3b872d18b973dea90cb9668c0ecbf0d53c64