[FIX] website_theme_install: restore correct external id for theme/hidden category
The website_theme_install needs to reference the module category whose
xml_id is base.module_category_theme_hidden. The problem was that it
needs to do this in a context where the category might not exist (as
auto-created only if a module defines its category as "Theme/Hidden").
To solve the problem, the category was explicitly created by the
website_theme_install module to be able to reference it.

Commit a08ac81c575628c01c8687576de8fe1a42b9927a removed the "base." part
of the xml_id thinking it was a record which had to be created by the
module. This then wrongly showed the Theme/Hidden modules in the list.

Now the solution came from commit https://github.com/odoo/odoo/commit/b57528fc#diff-07193f47475ccdd78f39773a368ce05bL386
which now allows to use the second parameter of ref in a domain
context. This allows to tell the system to not raise an exception if
the ref we want to find does not exist.