[FIX] website_project_issue: apply 0f35cd3 to count controller
The goal is to display the number of issues that are in projects
with a "portal" privacy setting. However, this setting is used by a
record rule to limit access to the project to followers only; this
means that you should add all the customers who submit issues to your
support project. This makes no sense.

To circumvent this problem, the domain of projects with this visibility
setting is fetched as sudo.