[FIX] website_sale: fix product website description translation
This is an oversight during the conversion to the new API
in the revision
bc7549e1e1fbf4a868b8ed57154355337f06357a

`html_translate` allows to split the source term
in several pieces, so the translations can be done
pieces by pieces.

For HTML fields, this is the default method of translation,
EXCEPT if `sanitize´ is set to `False`.
Unfortunately, this is the case for this field, and therefore
`html_translate` must still be precised for this field,
while, for other HTML fields, not having `sanitize` set to `False`,
`translate=True` is enough to use the method `html_translate`

opw-688470