[REV] fix `ir.qweb#render()` return type
This commit reverts the API change introduced in fabd61b398d1390591f3e6750039a5ccf57cc421
that changes `ir.qweb#render()` to return a unicode string instead of a utf8
encoded bytestring.

Returning a unicode string is undoubtedly cleaner and a step in the direction
of python 3 but still is an API change that can break third party modules.

Let's wait the migration to python 3 before breaking other people's toys.