[IMP] web_editor: mark dirty text to translate with the translated color
The content to translate is yellow and the translated content is green
in page translation mode. When the user starts translating a paragraph
it stayed yellow, making it difficult to see the remaining content to
translate when translating a big page.

This commit colors the dirty text in a paler green so that the user can
see the distinction.