Revert "[IMP] stock: Remove domain on source/destination on Operation Details"
This reverts commit dedc6085f9b64d7f96fff09f90f67df88defd258.

The domain makes clear that the location to be chosen in is among the children of the location zones chosen in the picking.  That way, if the stock would be calculated in the past (e.g. also for the stock valuation report), the system is going to think it is maybe still internal, while it went out, ..., so those domains are crucial.