[IMP] website_blog: add consistent behaviors for cover options
There was no bug but now the behaviors are more consistent :

- Demo data now declare a black invisible filter instead of nothing.
This allows user to see the effect of filter intensity option
(otherwise it had no visible effect since there was no filter color).

- If the filter intensity is set to "None" (invisible filter) and that
a filter color option is highlighted or chosen, a "Low" filter
intensity is automatically highlighted or chosen so that the user see
the filter color he is chosing (instead of having to choose the filter
intensity first).

- Hide the "Size" option on the blog list page since it has not visible
effect there.

- Some code "performance" improvements in option selection.