[FIX] website: load customize menu at once if already opened
Following commit 59c6344158f39d9f33606ca545d78ce99a0aebdf, the bug for customize menu
could have been ignored but as it is an easy fix :

The customize menu is filled after a rpc which is performed on
mousedown on the menu toggle. But in the case the user clicked on the
customize menu before the mousedown handler was attached... nothing
occured. It is possible since the customize menu is rendered but its
JS widget is only attached to it once everything is ready.

Now, the menu is filled on first mousedown or if the menu is already
opened when the JS widget is attached.