[FIX] website: fix library incompatibility for moves of menus
Since the library nestedSortable was updated to the release 2.1a,
at the below revision:
https://github.com/odoo/odoo/commit/5cd5bd3cb5cd626ec55cbcbee89f3ea979e7986f
It was no longer possible, in the front-end, to move the menus:
Re-sequence them, put one below another, ...

Indeed, in that piece of code, `.item_id` was never
set with the new release, while it was the case with the former one.
We replace with `.id`, which looks like to have the same behavior.