[FIX] website_blog: handle replies to Blog Post Published as notes
As the Blog Post Published message is posted on a channel every reply
to this message is also posted on the blog. However blogs may holds
lot of followers not necessarily interested in being spammed by answers
to that message. We consider now that those replies are notes to avoid
spam. Blog managers can follow the note subtype to be warned about
those messages.