[FIX] web_planner, account: add support for enterprise module + fix link 'Revenue Recognition'
Now, when the module is not found, instead to display a traceback, we open a popup to upgrade to Enterprise.

That will be wrong in case of outdated / wrong link, but in most cases, it is because it is an enterprise module.

This commit closes #13882

TODO: Remove zindex 1051 on planner in master/saas-14